Add files via upload
The goal of the project was to develop a text emotion detection system by combining a support vector machine (SVM) classifier with OpenAI GPT-3. The project aimed to accurately identify and analyze emotions in textual data by leveraging machine learning and natural language processing techniques. The integration of GPT-3 was used to generate sequences of sentences, and the SVM classifier was employed to predict emotions for these sequences. The ultimate objective was to determine the dominant emotion for a given text sequence based on the highest probability assigned by the classifier.